Start to add the infrastructure for resolving locations before queries
Backends can now indicate if they need resolution for a given location
via a new virtual method, and Manager will take care of this so we don't
have to duplicate the same logic in all backends.

So far only implement for departures, the same approach is still needed
for journeys too.