Fix IBNR and UIC station code usage as location identifiers
No backend supporting those is using them exclusively for various reasons
(a smaller station might not have such an id for example), so verify we
actually got an IBNR or UIC station code, and if not, namespace the
identifier accordingly.

This matters for location merging, and for KItinerary which can use those
ids to look up further information about a station in its database.