Add support for language selection in backends
There are backends where we need to request the language of user visible
messages explicitly (Hafas, EFA), rather than getting all possible
translations in the result. For this we now record in the backend service
configuration which languages are supported, and pick the best one of
those based on the current locale.