Compute z-indexed hash buckets and write out the corresponding data tables
This basically gives us a simple hash map using 2N leading bits of the
z-order curve position of N-sized bounding box tiles as key. N is chosen
so that no bounding box tile would contain two lines of the same name.
Or in other words: with this we can find a unique line meta data record
based on a geo coordinate along that line and its name.

This should in theory further benefit from compressing key ranges, however
for the current trial data the tile sizes are so large that we hardly ever
seem to have adjacent ones with the same content.