Augment line meta data by data from Wikidata
So far this is adding line colors when available, as well as line logos.

Even the small test set right now shows a number of inconsistencies between
OSM and Wikidata, so this will need more sanity checks in our code, as well
as upstream fixes in the data. The process as such seems to work though.