Allow to feed an OSM XML file directly into the code generator
Querying this via Overpass is not scaling very well when we want to extend
this to world-wide coverage. Instead, this is for exploring doing basic
filtering using osmfilter on a full local copy of the full OSM database,
and then doing the detailed processing on a sufficiently reduced dataset
in here.