Enforce a minimum distance between lines with the same name
This is set to not go below dropping at least 40bit of the z order
curve position for our hash table, to avoid this exploding on dubious
input data.

This identified two problematic lines in the input data set, the Bundang
Line in Seoul and the Green Line in Tokyo. Both seems to be only present
as individual segments, lacking the overall route_master relation, and are
positioned in a way that all our fallback mechanisms for re-grouping this
fail.