Turn spatial index for line metadata into a quad tree
Like the previous approach this is based on a z-order curve to build the
tiles, but now we are no longer constraint to a single tile size.
This increases the complexity of the generator a bit, but has no impact
on lookup performance and most importantly reduces the space needed to
store this to about 10% of the previous approach.