Also store mode in the line metadata
Allows more precise matching (given the frequently very ambiguous names),
and allows us to improve mode information if we only got a generic one
like "Line::Train" from the backend.

The extra storage space this needs for now is something we likely can get
rid of again with a more fine-granular memory layout, once we have the full
data set in and know the final space requirements of all parts.