Make better use of OSM links to product Wikidata items
Ideally OSM links to a Wikidata item representing a line, but that isn't
always the case and we get an item describing the mode, product or operator
instead. We can't do much with operator items, but we can feed the mode or
product ones in the existing product handling code.