Don't use a line logo as a candidate for a product logo
This can happen with sub-optimal types for the Wikidata item, and isn't
even wrong in all cases, such as there being no line items in Wikidata
and we have a reference to a product item, or a hybrid thereof. Therefore
we only remove the presumed line icons here if there are multiple
alternatives for the product icon.