Rename Departure to Stopover
Departure was always a poor name, given it is also used for arrivals.
Things got worse with this now also being used for intermediate stops,
which are both arrivals and departures at the same time. Finding a good
alternative name is tricky though, as many candidates are ambiguous
regarding stop locations ("stop", "halt", etc). Digitransit uses the
words "call" and "quay" to distinguish this, but while clear those are
rather niche terms. FPTF uses "stopover" for this, which seems like a
decent compromise.

This is only the first step in adjusting the entire API accordingly,
more classes will follow, while trying to keep source compatibility.