Ignore layers when their value matches the current level
This is technically wrong, but it seems to be a common issue in the OSM
data that layers are assigned globally rather than per level (which is
what the MapCSS rendering algorithm assumes).

It would be better to find a way to detect the presence of this problem
and only then apply this workaround, but until then this seems to produce
consistently better results.