Initial attempt on reusing of geometry between scene graph updates
For zoom changes this already entirely eliminates geometry creation and
their associated memory allocations, for level changes a similar
optimization is now possible.

The top costs on scene graph updates are now string-based tag lookup as
part of style evaluation and material-related allocations. Both should be
fixable. None of this should be relevant for scenes of the size this was
designed for (a single station or airport), it only starts to become
clearly visible in multi-million item full-city benchmark scenes.