Start of a zero-copy O5M reader
O5M is the binary format for OSM data, which is for example used by the
Marble vector tile server. So far this can just iterate over the file
structure and decode node coordinates.

This works on mmap'ed files, which should allow us to use the tag string
data directly from here. That however will still require larger changes
in the basic data structures that are currently too much focused on the
OSM XML parser needs.