Add automatic boundary searcher
Given a coordinate this finds the area of interest, such as the train
station or airport at this location. We now incrementally load map tiles
until we have that area fully covered.

This does basic geometry re-assembly from Marble vector tiles, which would
otherwise interfere massively with this. Once we have a generic solution
for that we can simplify things here. Beyond that this of course will
need more fine-tuning and polish to work properly in more cases.