Defer processing area fragments we cannot merge yet
This happens when a polygon intersects a tile in more than one area, those
we cannot merge without also having the connecting tile as well. So we need
to defer this until then, or if we never get the connecting tile because
that's out of range, we just keep the synthesized polygons.