Consider explicit and default units when taking line width from tag values
So far this assumed the tag value is a numeric value in meters. This
broke for values using an explicit unit (e.g. runway width in BRU specified
in feet), as well as for tags that have a different default unit defined
(e.g. railway track gauge, which is specified in millimeters).