Add infrastructure for showing dynamic overlays on top of the map
There's currently one source of overlays, in form of a generic QAIM
adapter, with rendering driven via MapCSS like the rest of the map.
We can use this for the existing gate/platform models and bring in
the arrival/departure locations that way, and we should be able to
bring in live data for vehicle positions via KPT::LocationQueryModel
as well with a bit of extra glue code eventually.