Make the MapCSS area selector cover multi-polygons as well
The MapCSS spec allows this but doesn't seem to require it. Apart from
the diagnostic stylesheet we never treat simple and complex polygons
differently in styling, so this makes things actually easier and less
error-prone for us, while separate treatment (should it ever be necessary)
is still possible by having a multi-polygon selector following a matching
area selector to specify the difference.