Allow to specify supported rental vehicle networks in OTP configuration
Unfortunately this cannot be queried online, but needs to be dug out of
server-side config files.

This enables a number of things though:
- filter static dummy networks that e.g. stadtnavi uses to provide
locations for taxis or car sharing.
- translate the internally used identifiers to human readable names.
- provide vehicle type default values for single-type networks, so we can
show appropriate icons etc.
- map identifiers to GBFS feeds where we have those available as well, to
support result merging.

We also still need to add this data to the other OTP instances supporting
rental vehicles (eg. those in Finland).