Add a tool to validate and pre-compute GBFS feed bounding boxes
We use GBFS's systems.csv as input for this, and generate a list of feeds
we ship from this. We can't use systems.csv directly for this as it doesn't
contain the bounding boxes, and we therefore would need to do a few hundred
HTTP GETs on first use, on every client, to find which feed is relevant for
a given area.