Add a basic GraphQL query minimizer
This reduces the amount of data we sent in requests by about 50% (prior
to compression). It also allows us to freely use comments now, without
leaking all of those to the server just to be discarded there.

In particular this allows for license annotations now, bringing us a step
closer to full REUSE compliance.