Port the remaining level filter code to use MapCSS
This adds a third filter state now for "dependent elements", ie. things
that make little sense on their own, but can be displayed if we have other
things on the same level as well. Underground railway lines are such an
example: They are important in the context of a subway station, but on
their own they add no value at all.