Beginning of a online test for all query types and backends
This isn't run as part of regular automatic tests given it takes quite some
time and produces tons of network requests, but it's nevertheless valuable
for easing continuous quality control on the various backend services.

The text matrix is still vastly incomplete, but this has already identified
a few issues.