Match free-standing platform sections against the corresponding platform
This helps with the case where platform section nodes aren't part of the
edge or area geometry. Checking whether the area geometry contains the
section node isn't good enough though, as we have to consider both the
case where we only have an edge or track geometry, and the case that the
section node is actually outside the area geometry (e.g. on the track).