Initial attempt of dealing with split edges/areas/tracks
This also fixes a pretty severe bug in Element::operator!= not working at
all so far. As a side-effect we get the (presumed) loading ramps in Gare
de Lyon detected now, which technically is correct, those a distinct
platforms. Before looking for a post-processing filter for those, let's
first see if we can fix that in the OSM data.