Let Stopover hold platform/vehicle layout information itself
This will allow providing those information directly from stopover or
journey queries. Most of this is internal for now, the old outward-facing
remains functional. The change to the cache format here is possible as
positive entries have a lifetime in the range of minutes only, so backward
compatibility solves itself quickly.