Add a group by country mode to BackendModel
This removes the restriction of a backend only being associated with a
single country. We work around that with aggregate country codes (EU, UN)
in a few places, but that isn't really ideal. Instead we can now show
backend entries for each country they are associated with.

While already more correct, the result isn't perfect yet regarding the
ordering, and filtering out fallback backends for countries where we have
better local sources is also still missing.

The previous flat view remains the default to not break existing users,
the new mode is however fully compatible with existing view that use the
country as section header already.