Produce valid results when simplifying areas close to the threshold
When the simplified area extends not significantly more than the size
threshold in one direction, the result of the Douglas Peucker algorithm
can deteriorate into a area-less polygon. In this case we fall back to
the bounding box instead now.

It's not that those small features have much practical impact, but the
resulting broken polygons cause problems in subsequent use.