Deal with colliding system ids
The specification requires those to be globally unique, but you guessed it,
they aren't... There's at least three different operators who seem to be
using a constant there, a UUID of all possible constants even in two cases.

We now identify those and assign our own system id based on a hashed feed
URL. This isn't ideal as the feed URL can change as new versions become
available, but it's the best we can do.