Start of OpenJourneyPlanner support
Only location queries work so far, the rest is mostly scaffolding still.

We need this as a replacement for the deactivated SBB Hafas endpoint. Due
to the close similarity of OJP and TRIAS we might also get support for
that as a side-effect. That in turn could help with the broken VVO EFA
endpoint.