More advanced OTP bounding polygon probing
This now additionally lists all stops and applies the same 3 sigma outlier
filter we also use for GBFS already, and intersects the result with the
officially reported bounding polygon. This seems to reliably fix outliers
for all affected OTP instances.

Helps to actually sit next to OTP developers/operators for a bit :)