Allow to specify access and egress modes in journey requests
That is, which individual transport options should be considered on the
first (access) and last (egress) leg of a journey. So far we always
implicitly assumed walking. Scenarios that are now also possible are e.g.
taking your bike and taking that with you on the train, or taking your car
to the station (and needing a place to park it there). Routing engines can
take all that into considerations, if we ask them to.

This is modeled based on OTP v2/Entur v3) but should cover the capabilities
of other engines as well.

This is only the request API so far though, support in any of the backends
is still missing, as is support for this in the result types.