Prefer backends containing both ends of a journey
We still want to use backends containing only one side as a fallback, as
those have a chance of working nevertheless (due to not detailed enough
coverage area data), but those containing both sides are certainly better.

Can reduce the amount of queries we do when no backend is explicitly
selected.