Initial onboard status API
This uses the Wifi SSID to detect whether we are on board of a train and
then queries the corresponding onboard API about the current position and
journey details.

This needs elevated access to Wifi details, only support for NM and
Android (partially) are implemented so far, a Flatpak containment wont
work.

In terms of supported providers, we have CD Railjet, DB ICE, SNCF INOUI
and Thalys so far, and this is extensible by adding a small script and a
few parameters.