Also look for backend configurations on disk
This helps during development, as you don't need to recompile to test
changes, and you can have development or demo/testing configurations
available without modifying the code.

A convenient way to do this is by symlinking src/lib/networks to
$XDG_DATA_DIRS/org.kde.kpublictransport/networs.

This could also come in handy for dynamic updates as discussed in #29.