Add support for Hafas rtMode TripSearch configuration parameter
This is only known to be used by DB, and it needs to be set explicitly
to "HYBRID" there to produce the expected results when faced with
replacement trains and/or cancellations.

This fix is brought to you by the current railway strike in Germany.

(cherry picked from commit 6b613a308be1e3003ab13d8635aff955e2849064)