Update Navitia coverage to match the current reality
A lot of areas have been removed unfortunately. But more importantly this
was outperformed by ÖBB when trying to find the "best" backend for France.
That however only has long-distance travel, so we were back to getting
really bizarre connections between the major train stations in Paris for
example.

With this now both are considered equal for queries in France, which
produces more sensible results.