Be a bit more flexible regarding modes when matching line metadata
There's systems in certain cities with massive ambiguities regarding
whether they are a rapid transit, a subway/metro or a tram. Accept
anything if the base mode is the same (rail, road, etc), as long as
the name is sufficiently highly identifying (that is, not a single
letter or number).

Seems to work as a compromise and produces much better results in e.g.
Stuttgart.