Add dedicated mode options for direct connections
That is, connections not involving public transport at all. Motis v2
provides this (and most other backends do as well implicitly on short
distances).

This avoids ambiguities with access/egress options which so far had been
abused for this, but that makes it impossible to distinguish between a
direct connection with your own bike and taking your bike on public
transport.