Add support for DB's proxied Hafas API
Mostly motivated by rumors of the imminent shutdown of the "real" Hafas
endpoint. There's a few information we are not getting this way anymore
(e.g. stopover coordinates, shapes), so for now this is mainly meant as
an emergency backup option.

We keep using the old Hafas API for now, in a local build this is available
as a second backend for testing.

(cherry picked from commit 6c1801afd5e6af8897d4fa3f036a678549119fbd)