Remove Location::identifiers
This is only used internally and for the examples, and leaks implementation
details about the identifier set into the API. Instead, expose a list of
available identifier types for the needs of the examples. As a bonus this
allows us to move the examples closer to being QML only.

This is a preparation for refactoring the identifier set code for reuse in
JourneySection.