Add trip query API
That is, querying one specific run along a route, based on a previously
queried JourneySection.

Backends having dedicated support for this will provide more efficient
results (as well as the full trip), for backends not having this this
is emulated by a journey query (which is less efficient and only returns
the queried potentially partial trip).