Fix MOTIS v2 admin area parsing
Due to a typo this was skipped entirely, with our location-based post-
processing hiding most of this. Now that it is used, also improve the
admin level selection a bit, but to work properly this still needs a
per-country mapping table.

(cherry picked from commit 5ea93bd28eb45676b2caddf1148ce85d3651d047)