Add geographic country bounding box data for biasing geocoding
Generated from OSM boundaries and automatically injected in LocationRequest
when having a country but no manual viewbox area set.

As the MOTIS backend already consumes this data when present this has the
immediate effect that searching for e.g. "Berlin" in DE or US gives you
the respective expected results.

Within a continent the results aren't as good yet, as we don't adjust the
bias for area size yet, and because the sorting in the location search
result model isn't taking any of that into account yet.