Consider supported location types when selecting geocoding backends
This removes an old hack that dealt with this exploiting the fact that
the two special backends for GBFS and accessibility.cloud can be identified
by their global coverage.

Instead backends now explicitly state which types they support, and we
keep looking for backends until all requested types are covered. In return
backends don't have to check requests for supported types anymore.